+++ server/src/main/java/server/util/QuestionConversion.java
@@ -1,8 +1,8 @@
 package server.util;
 
 import commons.CompareQuestion;
-import commons.Question;
 import commons.Activity;
+import commons.Question;
 import commons.WattageQuestion;
 import java.util.Random;
 
@@ -18,7 +18,7 @@ public class QuestionConversion {
         Question result = null;
         if (activities != null && activities.length == 4){
             String[] titles = new String[4];
-            int[] wattages = new int[4];
+            long[] wattages = new long[4];
             for (int i = 0; i < 4; i++){
                 if (activities[i] != null) {
                     Activity current = activities[i];
